+++ widgets/src/buttons/fab_button.rs
@@ -74,7 +74,7 @@ impl ComposeDecorator for FabButtonDecorator {
 /// ```
 #[derive(Default, Declare)]
 pub struct FabButton {
-  #[declare(default=Palette::of(ctx!()).primary())]
+  #[declare(default=Palette::of(BuildCtx::get()).primary())]
   color: Color,
 }